+++ src/slice.c
@@ -8,6 +8,7 @@ SEXP fns_vec_slice_dispatch = NULL;
 // Defined below
 SEXP vec_as_index(SEXP i, SEXP x);
 static void slice_names(SEXP x, SEXP to, SEXP index);
+static SEXP vec_slice_impl(SEXP x, SEXP index, bool dispatch);
 
 
 static void stop_bad_index_length(R_len_t data_n, R_len_t i) {
@@ -76,51 +77,81 @@ static SEXP list_slice(SEXP x, SEXP index) {
 
 #undef SLICE_BARRIER
 
+static SEXP df_slice(SEXP x, SEXP index) {
+  R_len_t n = Rf_length(x);
+  SEXP out = PROTECT(Rf_allocVector(VECSXP, n));
+
+  SEXP nms = Rf_getAttrib(x, R_NamesSymbol);
+  Rf_setAttrib(out, R_NamesSymbol, nms);
+
+  for (R_len_t i = 0; i < n; ++i) {
+    SEXP sliced = vec_slice_impl(VECTOR_ELT(x, i), index, true);
+    SET_VECTOR_ELT(out, i, sliced);
+  }
+
+  UNPROTECT(1);
+  return out;
+}
+
+
 static SEXP vec_slice_dispatch(SEXP x, SEXP index) {
   return vctrs_dispatch2(syms_vec_slice_dispatch, fns_vec_slice_dispatch,
                          syms_x, x,
                          syms_i, index);
 }
 
-static SEXP vec_slice_switch(SEXP x, SEXP index, bool dispatch) {
+static SEXP vec_slice_impl(SEXP x, SEXP index, bool dispatch) {
+  if (has_dim(x)) {
+    return vec_slice_dispatch(x, index);
+  }
+
+  SEXP out = R_NilValue;
+
   switch (vec_typeof_impl(x, dispatch)) {
   case vctrs_type_null:
     Rf_error("Internal error: Unexpected `NULL` in `vec_slice_impl()`.");
 
   case vctrs_type_logical:
-    return lgl_slice(x, index);
+    out = lgl_slice(x, index);
+    break;
   case vctrs_type_integer:
-    return int_slice(x, index);
+    out = int_slice(x, index);
+    break;
   case vctrs_type_double:
-    return dbl_slice(x, index);
+    out = dbl_slice(x, index);
+    break;
   case vctrs_type_complex:
-    return cpl_slice(x, index);
+    out = cpl_slice(x, index);
+    break;
   case vctrs_type_character:
-    return chr_slice(x, index);
+    out = chr_slice(x, index);
+    break;
   case vctrs_type_raw:
-    return raw_slice(x, index);
+    out = raw_slice(x, index);
+    break;
   case vctrs_type_list:
-    return list_slice(x, index);
+    out = list_slice(x, index);
+    break;
+
+  case vctrs_type_dataframe:
+    out = PROTECT(df_slice(x, index));
+    out = df_restore(out, x, index);
+    UNPROTECT(1);
+    return out;
 
   default:
-    return R_NilValue;
-  }
-}
-
-static SEXP vec_slice_impl(SEXP x, SEXP index, bool dispatch) {
-  SEXP out = R_NilValue;
-
-  if (!has_dim(x)) {
-    out = vec_slice_switch(x, index, dispatch);
-  }
-  if (out == R_NilValue) {
-    return vec_slice_dispatch(x, index);
+    out = PROTECT(vec_slice_dispatch(x, index));
+    out = vec_restore(out, x, index);
+    UNPROTECT(1);
+    return out;
   }
 
   PROTECT(out);
   slice_names(out, x, index);
-  UNPROTECT(1);
 
+  out = vec_restore(out, x, index);
+
+  UNPROTECT(1);
   return out;
 }