+++ src/vctrs.h
@@ -132,13 +132,16 @@ extern Rcomplex vctrs_shared_na_cpl;
 R_len_t vec_size(SEXP x);
 SEXP vec_cast(SEXP x, SEXP to);
 SEXP vec_slice(SEXP x, SEXP index);
+SEXP vec_restore(SEXP x, SEXP to, SEXP i);
 
 bool is_data_frame(SEXP x);
 bool is_record(SEXP x);
 bool is_scalar(SEXP x);
 
+R_len_t df_size(SEXP x);
 R_len_t df_rownames_size(SEXP x);
 R_len_t df_raw_size(SEXP x);
+SEXP df_restore(SEXP x, SEXP to, SEXP i);
 
 // Most vector predicates return `int` because missing values are
 // propagated as `NA_LOGICAL`