+++ tests/testthat/test-cast.R
@@ -176,7 +176,7 @@ test_that("data frame vec_restore() checks type", {
 
 test_that("can use vctrs primitives from vec_restore() without inflooping", {
   scoped_global_bindings(
-    vec_restore.vctrs_foobar = function(x, to) {
+    vec_restore.vctrs_foobar = function(x, to, ...) {
       vec_type(x)
       vec_na(x)
       vec_assert(x)
@@ -187,4 +187,11 @@ test_that("can use vctrs primitives from vec_restore() without inflooping", {
 
   foobar <- new_vctr(1:3, class = "vctrs_foobar")
   expect_identical(vec_slice(foobar, 2), "woot")
+})
+
+test_that("vec_restore() passes `i` argument to methods", {
+  scoped_global_bindings(
+    vec_restore.vctrs_foobar = function(x, to, ..., i) i
+  )
+  expect_identical(vec_slice(foobar(1:3), 2), 2L)
 })